backend/requirements.txt

fastapi
uvicorn
requests
pydantic-ai
python-dotenv
httpx
groq


